Resolving duplicate code block issue
