+++ packages/azure-functions/ProcessTrustedFile/helpers/process-metric-upload.js
@@ -1,7 +1,6 @@
 import buildSignalRMessage from '../../Shared/build-signalr-message.js'
 import bngMetricService from '@defra/bng-metric-service'
 import { blobStorageConnector } from '@defra/bng-connectors-lib'
-import processMetric from '../../Shared/process-metric-data.js'
 
 export default async function (context, config) {
   let signalRMessageArguments, metricData
@@ -25,7 +24,6 @@ export default async function (context, config) {
       throw new Error('Unable to retrieve blob')
     }
 
-    console.log('metricData==>', metricData)
     signalRMessageArguments = [{
       location: config.fileConfig.fileLocation,
       metricData: processMetric(metricData)
@@ -37,3 +35,17 @@ export default async function (context, config) {
     context.bindings.signalRMessages = [buildSignalRMessage(config.signalRMessageConfig, signalRMessageArguments)]
   }
 }
+
+const processMetric = metricData => ({
+  startPage: metricData.start,
+  d1: metricData.d1OffSiteHabitatBaseline,
+  d2: metricData.d2OffSiteHabitatCreation,
+  d3: metricData.d3OffSiteHabitatEnhancement,
+  e1: metricData.e1OffSiteHedgeBaseline,
+  e2: metricData.e2OffSiteHedgeCreation,
+  e3: metricData.e3OffSiteHedgeEnhancement,
+  f1: metricData.f1OffSiteWaterCBaseline,
+  f2: metricData.f2OffSiteWaterCCreation,
+  f3: metricData.f3OffSiteWaterCEnhancement,
+  validation: metricData.validation
+})
